Add ttf support to sdl wrapper and render simple fps counter to window
This commit adds an SDL2_ttf wrapper. With this you are now able to
create textures that contain fonts and render them to the screen. At the
moment there is also a crude FPS counter being rendered on the screen.
However, this counter is highly inefficient and will need a rework since
at the moment it takes an additional 5% CPU power to compute.