Add sdl2 cmake find files as git submodule
This commit adds a git submodule that searches your system for the
necessary SDL2 libraries. This is done so that we are not adding the
gcc library linkers in manually, this allows us to be a bit more
flexible when it comes to compiling across different systems.

Reference: https://github.com/aminosbh/falling-brick-game/tree/master/cmake/sdl2